Fixed a number of different bugs in the GPU code, particularly sync of buffers

+++ Include/Geometry/strided_reads_gpu.h
@@ -23,14 +23,14 @@ enum DIRECTION { UP = 0, DOWN = 1 };
 #ifdef __cplusplus
 
 template <typename REAL, typename FIELD_TYPE, typename SITE_TYPE>
-__host__ __device__ void read_gpu(int stride, SITE_TYPE *s, const FIELD_TYPE *in, int ix, int comp, int dim) {
+__host__ __device__ void read_gpu(int stride, SITE_TYPE *s, const FIELD_TYPE *in, size_t ix, int comp, int dim) {
     const int field_dim = sizeof(FIELD_TYPE) / sizeof(REAL);
     const int n_components = sizeof(SITE_TYPE) / sizeof(REAL);
 #ifdef FIXED_STRIDE
-    int iz = ((ix / THREADSIZE) * THREADSIZE) * dim * field_dim + (ix % THREADSIZE) + ((comp)*n_components) * (THREADSIZE);
+    size_t iz = ((ix / THREADSIZE) * THREADSIZE) * dim * field_dim + (ix % THREADSIZE) + ((comp)*n_components) * (THREADSIZE);
     const int _stride = THREADSIZE;
 #else
-    int iz = ix + ((comp)*n_components) * (THREADSIZE);
+    size_t iz = ix + ((comp)*n_components) * (THREADSIZE);
     const int _stride = stride;
 #endif
     REAL *in_cpx = (REAL *)in;
@@ -42,14 +42,14 @@ __host__ __device__ void read_gpu(int stride, SITE_TYPE *s, const FIELD_TYPE *in
 }
 
 template <typename REAL, typename FIELD_TYPE, typename SITE_TYPE>
-__host__ __device__ void write_gpu(int stride, SITE_TYPE *s, FIELD_TYPE *out, int ix, int comp, int dim) {
+__host__ __device__ void write_gpu(int stride, SITE_TYPE *s, FIELD_TYPE *out, size_t ix, int comp, int dim) {
     const int field_dim = sizeof(FIELD_TYPE) / sizeof(REAL);
     const int n_components = sizeof(SITE_TYPE) / sizeof(REAL);
 #ifdef FIXED_STRIDE
-    int iz = ((ix / THREADSIZE) * THREADSIZE) * dim * field_dim + (ix % THREADSIZE) + (comp)*n_components * (THREADSIZE);
+    size_t iz = ((ix / THREADSIZE) * THREADSIZE) * dim * field_dim + (ix % THREADSIZE) + (comp)*n_components * (THREADSIZE);
     const int _stride = THREADSIZE;
 #else
-    int iz = ix + ((comp)*n_components) * (THREADSIZE);
+    size_t iz = ix + ((comp)*n_components) * (THREADSIZE);
     const int _stride = stride;
 #endif
     REAL *out_cpx = (REAL *)out;
